# Gamma optimizations and benchmarks