--- a/.dir-locals.el
+++ b/.dir-locals.el
@@ -1,7 +1,7 @@
 ((nil
   . ((cider-default-cljs-repl . node)
      (cider-preferred-build-tool . clojure-cli)
-     (cider-clojure-cli-aliases . ":test:nextjournal/clerk")
+     (cider-clojure-cli-aliases . ":benchmark:test:nextjournal/clerk")
 
      ;; Custom indentation:
      (eval . (put-clojure-indent 'gen 1)))))