--- a/bb.edn
+++ b/bb.edn
@@ -27,6 +27,10 @@
   {:doc "Run CLJS tests."
    :task (shell "npm run test")}
 
+  benchmark:clj
+  {:doc "Run CLJ benchmarks."
+   :task (shell "clojure -M:benchmark")}
+
   build-static
   {:doc "Generate a fresh static build."
    :task