Turns out that pre-computing constants and type-hinting functions has a big performance gain:

- `(gamma 2.5)` 77.40% speedup (execution mean 4.889986 µs → 1.104972 µs)
- `(log-gamma 10)` 63.92% speedup (execution mean 4.845715 µs → 1.748155 µs)
- `(log-gamma-1p -0.05)` 76.86% speedup (execution mean 4.981068 µs → 1.152443 µs)
- `(inv-gamma-1pm1 -0.05)` 77.85% speedup (execution mean 4.582663 µs → 1.015241 µs)
- `(lanczos-approximation 170)` 73.15% speedup (execution 4.276010 µs → 1.148002 µs)

Did this for `math.gamma` and `math.utils`. Also experimenting with a way to define and run benchmarks like tests: Benchmarks are defined in the `/benchmark` directory, then functions to benchmark are tagged with `^:benchmark` and profiled with Criterium. Just run `bb benchmark:clj` to compute benchmarks! See `gamma_benchmark.cljc` for an example.