@eightysteele 2024-05-14
@sritchie yep, i responded—lmk if that explanation about `:benchmark` metadata makes sense!